CODE RUB : Separate logic test for Remove assignments
